Moved models to separate folder + MasterDetail
#403
