+++ templates/Pages/MasterDetail.CodeBehind/.template.config/template.json
@@ -14,7 +14,8 @@
     "wts.type": "page",
     "wts.framework": "CodeBehind",
     "wts.version": "1.0.0",
-    "wts.order": "1"
+    "wts.order": "1",
+    "wts.dependencies": "wts.Feat.SampleDataService"
   },
   "sourceName": "MasterDetail",
   "preferNameDirectory": true,
@@ -36,13 +37,7 @@
     },
     {
       "path": ".\\Views\\MasterDetailDetailControl.xaml.cs"
-    },
-    {
-      "path": ".\\Models\\SampleModel.cs"
-    },
-    {
-      "path": ".\\Services\\SampleModelService.cs"
-    }    
+    }   
   ],
   "symbols": {
     "wts.rootNamespace": {
