+++ templates/Pages/MasterDetail/.template.config/template.json
@@ -14,7 +14,8 @@
     "wts.type": "page",
     "wts.framework": "MVVMBasic|MVVMLight",
     "wts.version": "1.0.0",
-    "wts.order": "1"
+    "wts.order": "1",
+    "wts.dependencies": "wts.Feat.SampleDataService"
   },
   "sourceName": "MasterDetail",
   "preferNameDirectory": true,
@@ -42,13 +43,7 @@
     },
     {
       "path": ".\\ViewModels\\MasterDetailDetailViewModel.cs"
-    },
-    {
-      "path": ".\\Models\\SampleModel.cs"
-    },
-    {
-      "path": ".\\Services\\SampleModelService.cs"
-    }    
+    }
   ],
   "symbols": {
     "wts.rootNamespace": {
