+++ templates/_composition/_shared/Page.Chart.SampleDataService/Models/DataPoint.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.ObjectModel;
+
+namespace Param_ItemNamespace.Models
+{
+    // TODO UWPTemplates: This is used by the Sample Chart Data. Remove this once your chart page is displaying real data
+    public class DataPoint
+    {
+        public double Value { get; set; }
+        public string Category { get; set; }
+    }
+}
