+++ templates/_composition/_shared/Page.Grid.SampleDataService/.template.config/template.json
@@ -13,6 +13,11 @@
  },
   "sourceName": "wts.ItemName",
   "preferNameDirectory": true,
+  "PrimaryOutputs": [
+    {
+      "path": ".\\Models\\Order.cs"
+    }
+  ],
   "symbols": {
     "wts.rootNamespace": {
       "type": "parameter",
