+++ templates/_composition/_shared/Page.Grid.SampleDataService/Services/SampleDataService_postaction.cs
@@ -107,16 +107,6 @@ namespace Param_ItemNamespace.Services
 
             return data;
         }
-
-        public class Order
-        {
-            public long OrderId { get; set; }
-            public DateTime OrderDate { get; set; }
-            public string Company { get; set; }
-            public string ShipTo { get; set; }
-            public double OrderTotal { get; set; }
-            public string Status { get; set; }
-        }
 //}]}
 
     }
