+++ templates/_composition/_shared/Page.MasterDetail.SampleDataService/Models/SampleModel.cs
